init commit change some address

+++ src/main/java/ir/rahgozin/wallet/application/wallet/command/CreditAccountCommand.java
@@ -2,10 +2,11 @@ package ir.rahgozin.wallet.application.wallet.command;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreditAccountCommand {
     private String requestId;
     private String accountNumber;
-    private Double amount;
-
+    private BigDecimal amount;
 }
